Moving all WP_Ajax functionality to the AjaxHandler class
